Refactor: Split out diffViewer modules
